requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-community==0.3.5
langchain-pinecone==0.2.0
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
pinecone-client==5.0.1
pypdf==5.1.0
python-dotenv==1.0.1
google-generativeai==0.8.3
scikit-learn==1.5.2
